Add files via upload
The aim of this project is to show the best 20 places in Alex on the map. So firstly, I scraped the data from Foursquare API. Converted it to a pandas data frame after that. Then, I cleaned and filtered the data. Finally, to map the data, I used folium (plotting library).