Add files via upload
The project's goal is to determine the relationship between the budget and other aspects. and then visualize the data using seaborn and matplotlip. So, first and foremost, I will thoroughly comprehend the data. After that, I'll clean the data and do some data exploration. Finally, look for correlations and visualize the results.